+++ tests/unit/test_models/test_full_battery_models/test_base_battery_model.py
@@ -337,6 +337,10 @@ class TestBaseBatteryModel(unittest.TestCase):
                 }
             )
 
+        # phases
+        with self.assertRaisesRegex(pybamm.OptionError, "multiple particle phases"):
+            pybamm.BaseBatteryModel({"particle phases": "2", "surface form": "false"})
+
     def test_build_twice(self):
         model = pybamm.lithium_ion.SPM()  # need to pick a model to set vars and build
         with self.assertRaisesRegex(pybamm.ModelError, "Model already built"):
@@ -386,6 +390,24 @@ class TestOptions(unittest.TestCase):
         # something that is the same in both domains
         self.assertEqual(options.negative["thermal"], "isothermal")
         self.assertEqual(options.positive["thermal"], "isothermal")
+
+    def test_domain_phase_options(self):
+        options = BatteryModelOptions(
+            {"particle mechanics": (("swelling only", "swelling and cracking"), "none")}
+        )
+        self.assertEqual(
+            options.negative["particle mechanics"],
+            ("swelling only", "swelling and cracking"),
+        )
+        self.assertEqual(
+            options.negative.primary["particle mechanics"], "swelling only"
+        )
+        self.assertEqual(
+            options.negative.secondary["particle mechanics"], "swelling and cracking"
+        )
+        self.assertEqual(options.positive["particle mechanics"], "none")
+        self.assertEqual(options.positive.primary["particle mechanics"], "none")
+        self.assertEqual(options.positive.secondary["particle mechanics"], "none")
 
 
 if __name__ == "__main__":